requirements.txt
absl-py==0.10.0
albumentations==0.4.6
apache-beam==2.23.0
argon2-cffi==20.1.0
astunparse==1.6.3
attrs==20.1.0
avro-python3==1.9.1
backcall==0.2.0
bleach==3.1.5
branca==0.4.1
Brotli==1.0.7
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
contextlib2==0.6.0.post1
crcmod==1.7
click==7.1.2
cycler==0.10.0
Cython==0.29.21
dash==1.13.4
dash-bootstrap-components==0.10.3
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
dnspython==1.16.0
dm-tree==0.1.5
docopt==0.6.2
easydict==1.9
entrypoints==0.3
et-xmlfile==1.0.1
fastavro==0.23.6
Flask==1.1.2
Flask-Compress==1.5.0
folium==0.11.0
future==0.18.2
gast==0.3.3
gin-config==0.3.0
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.27.2
google-cloud-core==1.4.1
google-crc32c==0.1.0
google-pasta==0.2.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
h5py==2.10.0
hdfs==2.5.8
httplib2==0.17.4
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
kaggle==1.5.6
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
kafka-python==2.0.1
kaleido==0.0.3.post1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mock==2.0.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.5
notebook==6.1.3
numpy==1.18.5
oauth2client==3.0.0
oauthlib==3.1.0
opencv-python==4.1.1.26
opencv-python-headless==4.4.0.42
openpyxl==3.0.4
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
plotly==4.9.0
promise==2.3
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
psutil==5.7.2
py-cpuinfo==7.0.0
pyarrow==0.17.1
pycocotools==2.0.1
pycparser==2.20
pydot==1.4.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pymongo==3.11.0
python-slugify==4.0.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.1
sentencepiece==0.1.91
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-addons==0.11.1
tensorflow-datasets==3.2.1
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
tensorflow-metadata==0.23.0
tensorflow-model-optimization==0.4.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==2.3.0.dev2020062301
text-unidecode==1.3
tf-models-official==2.3.0
tf-slim==1.1.0
tifffile==2020.8.25
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
typeguard==2.9.1
typing-extensions==3.7.4.3
urllib3==1.24.3
uritemplate==3.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
zipp==3.1.0
